requirements.txt
pyyaml~=6.0
tqdm~=4.64.1
opencv-python~=4.6.0.66
matplotlib~=3.6.2
numpy~=1.24.1
pandas~=1.5.2
seaborn~=0.12.1
pillow~=8.2.0
torchvision~=0.13.1
requests~=2.28.1
flask~=2.2.2
thop~=0.1.1-2209072238
einops~=0.6.0